Don't send google tracking on dev
Continue using `npm run dev` for local development.

To test tracking locally, use `npm run:tracking`

The weird argument passaround in server.js was needed because `process`
in index.js is different on client from server, and there didn't seem to
be a simple way to pass CLI arguments to the client, even with
`componentWillMount`, which is supposedly only ever called for server
rendering.

Weird but working is better than not working, I guess?!